# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "thaeeithantun.dev@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
...
        "/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get list of blogs",
                "description": "Returns list of blogs",
                "operationId": "get Blogs ",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "blogs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Blog Id",
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "description": "Blog Title",
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "description": "Blog Description",
                                                        "type": "string"
                                                    },
                                                    "created_by": {
                                                        "description": "Blog creater's name",
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "description": "Blog Created at",
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "description": "Blog Updated at",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "collectionFormat": "multi"
                                        },
                                        "message": {
                                            "description": "Message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Create blog with form data",
                "operationId": "createblogWithForm",
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Enter Blog Title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Enter Blog Description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Blog Id",
                                            "type": "integer"
                                        },
                                        "title": {
                                            "description": "Blog Title",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Blog Description",
                                            "type": "string"
                                        },
                                        "created_by": {
                                            "description": "Blog creater's name",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Blog Created at",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "Blog Updated at",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/blogs/{blogId}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get blog information",
                "description": "Returns blog data",
                "operationId": "getblogById",
                "parameters": [
                    {
                        "name": "blogId",
                        "in": "path",
                        "description": "Blog Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Blog Id",
                                            "type": "integer"
                                        },
                                        "title": {
                                            "description": "Blog Title",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Blog Description",
                                            "type": "string"
                                        },
                                        "created_by": {
                                            "description": "Blog creater's name",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Blog Created at",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "Blog Updated at",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Updates blog with form data",
                "operationId": "updateblogWithForm",
                "parameters": [
                    {
                        "name": "blogId",
                        "in": "path",
                        "description": "ID of blog that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Updated title of the blog",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated description of the blog",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Delete blog",
                "description": "Delete blog",
                "operationId": "deleteblogById",
                "parameters": [
                    {
                        "name": "blogId",
                        "in": "path",
                        "description": "Blog Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "security": [
        {
            "sanctum": []
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}